refactor: move plausible change request state to context file
