+++ frontend/src/component/changeRequest/ChangeRequestContext.tsx
@@ -1,8 +1,16 @@
 import { createContext, useContext } from 'react';
+import { ChangeRequestState } from './changeRequest.types';
+
+export type PlausibleChangeRequestState =
+    | Exclude<ChangeRequestState, 'Scheduled'>
+    | 'Scheduled pending'
+    | 'Scheduled failed'
+    | 'Scheduled suspended';
 
 const defaultContext = {
     willOverwriteStrategyChanges: false,
     registerWillOverwriteStrategyChanges: () => {},
+    previousState: 'Draft' as PlausibleChangeRequestState,
 };
 
 const ChangeRequestPlausibleContext = createContext(defaultContext);
