+++ frontend/src/component/changeRequest/changeRequest.types.ts
@@ -100,12 +100,6 @@ export interface IChangeRequestChangeBase {
     };
 }
 
-export type PlausibleChangeRequestState =
-    | Exclude<ChangeRequestState, 'Scheduled'>
-    | 'Scheduled pending'
-    | 'Scheduled failed'
-    | 'Scheduled suspended';
-
 export type ChangeRequestState =
     | 'Draft'
     | 'Approved'
