+++ frontend/src/hooks/api/actions/useChangeRequestApi/useChangeRequestApi.ts
@@ -1,9 +1,11 @@
 import useAPI from '../useApi/useApi';
 import { usePlausibleTracker } from '../../../usePlausibleTracker';
-import { PlausibleChangeRequestState } from 'component/changeRequest/changeRequest.types';
 import { getUniqueChangeRequestId } from 'utils/unique-change-request-id';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
-import { useChangeRequestPlausibleContext } from 'component/changeRequest/ChangeRequestContext';
+import {
+    PlausibleChangeRequestState,
+    useChangeRequestPlausibleContext,
+} from 'component/changeRequest/ChangeRequestContext';
 
 export interface IChangeSchema {
     feature: string | null;
